streamlit_app/server/server_requirements.txt
accelerate==1.1.1
faiss-cpu==1.9.0.post1
fastapi==0.115.5
huggingface-hub==0.26.3
joblib==1.4.2
lightgbm==4.5.0
llvmlite==0.43.0
numba==0.60.0
optuna==4.1.0
packaging==24.2
pandas==2.2.3
psutil==6.1.0
pydantic==2.10.2
pydantic_core==2.27.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sympy==1.13.3
torch==2.3.0
tqdm==4.67.1
transformers==4.46.3
triton==2.3.0
umap-learn==0.5.7
urllib3==2.2.3
uvicorn==0.32.1